backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
bcrypt==4.0.1
